[Foundation] Move the NSUrlSessionHandlerDelegate out of the handler file.
